Set maxSockets for nano
This will ensure the amount of open handles does not go over the limit, which is lower on macOS than on Linux.
Because all the tests are running in parallel a lot of sockets are opened to the DB. The intention with this option is that any sockets over the maximum will enter a queue.